// FrontEnd/src/app/services/models/user-front-end-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Address } from '../models/address';
export interface UserFrontEndResponse {
  address?: Address;
  email?: string;
  id?: number;
  url_back?: string;
  url_picture?: string;
  username?: string;
}